# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Medidor de Eficiência
  description: >
    Sistema web para registro de atividades diárias e análise de eficiência pessoal.
    A aplicação é um MVC server-side renderizado (Express + EJS): a maioria das rotas
    responde com HTML e usa redirects (302) após operações de escrita, seguindo o
    padrão Post/Redirect/Get de formulários HTML tradicionais — não é uma API JSON.
    Esta spec documenta a superfície de rotas para referência de desenvolvimento.
  version: "1.0.0"
servers:
  - url: http://localhost:3000
    description: Ambiente de desenvolvimento
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: >
        Sessão autenticada via express-session (RNF02). Rotas privadas sem sessão
        válida redirecionam (302) para /login.
...
paths:
  /login:
    get:
      tags: [Auth]
      summary: Exibe o formulário de login
      responses:
        "200":
          description: Página de login renderizada (HTML)
          content:
            text/html: {}
        "302":
          description: Usuário já autenticado — redireciona para a página inicial (/)
    post:
      tags: [Auth]
      summary: Autentica com e-mail e senha
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [email, senha]
              properties:
                email:
                  type: string
                  format: email
                senha:
                  type: string
      responses:
        "302":
          description: >
            Sucesso: redireciona para / com sessão criada (Set-Cookie).
            Falha (credenciais inválidas ou payload inválido): redireciona de volta para /login
            com mensagem de erro via flash.
          headers:
            Location:
              schema:
                type: string
                example: /
  /register:
    get:
      tags: [Auth]
      summary: Exibe o formulário de cadastro
      responses:
        "200":
          description: Página de cadastro renderizada (HTML)
          content:
            text/html: {}
        "302":
          description: Usuário já autenticado — redireciona para a página inicial (/)
    post:
      tags: [Auth]
      summary: Cria uma conta local (regra 1 — senha mínima de 8 caracteres, hash bcrypt)
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [nome, email, senha]
              properties:
                nome:
                  type: string
                  minLength: 2
                email:
                  type: string
                  format: email
                senha:
                  type: string
                  minLength: 8
      responses:
        "302":
          description: >
            Sucesso: cria a conta, autentica e redireciona para /.
            Falha (e-mail já cadastrado ou payload inválido): redireciona de volta para /register
            com mensagem de erro via flash.
...